move load function to module init
